RL used to solve the 4x4 Forzen lake problem
Added video for untrained and trained agent. Use RL algorithm to train the agent and also, display some basic information.
Main thing is the approach used to solve the problem at hand.

You will see how to use the EPSILON decay and how to use the Learning rate and Gamma.

We will also see how the next_optimal_q_val is deduced and how the next_q val is found. How to update the q_table.

And once the training is done, how the new Q_Table is used to play the game. 

ENJOY the video !